Work on directives for PGI

+++ simpeg_drivers/components/models.py
@@ -588,7 +588,7 @@ class InversionModel:
         """
         if isinstance(model, NumericData):
             model = self.obj_2_mesh(model, self.driver.inversion_mesh.entity)
-            model = self.driver.inversion_mesh.permutation @ model
+            model = (self.driver.inversion_mesh.permutation @ model).astype(model.dtype)
         else:
             nc = self.driver.inversion_mesh.mesh.n_cells
             if isinstance(model, int | float):
